MVP/app/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
bcrypt==4.3.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
aio-pika==9.3.1
pika==1.3.2
jinja2==3.1.3
loguru==0.7.2
pydantic==2.6.4
pydantic[email]
httpx==0.27.0
pytest==8.1.1
pytest-asyncio==0.23.6
# numpy==1.26.4
# pandas==2.2.2
# scikit-learn==1.5.0
